Feature/load relation with histogram

+++ src/statistics.cpp
@@ -4,17 +4,17 @@
 
 using namespace::std;
 
-Histogram::Histogram(unsigned int interval_width) {
+Histogram::Histogram(uint64_t interval_width) {
     this->interval_width = interval_width;
     this->interval_count = vector<size_t> (0, 0);
 }
 
-Histogram::Histogram(unsigned int interval_width, unsigned int estimated_histogram_max) {
+Histogram::Histogram(uint64_t interval_width, uint64_t estimated_histogram_max) {
     this->interval_width = interval_width;
     this->interval_count = vector<size_t> (get_interval_index(estimated_histogram_max) + 1, 0);
 }
 
-void Histogram::add_entry(unsigned int entry) {
+void Histogram::add_entry(uint64_t entry) {
     size_t i_index = get_interval_index(entry);
     if (i_index + 1 > get_number_of_intervals()) {
         interval_count.resize(i_index + 1, 0);
@@ -33,13 +33,13 @@ size_t Histogram::get_total_number_of_records() {
     return count;
 }
 
-size_t Histogram::get_number_of_records_geq(unsigned int threshold) {
+size_t Histogram::get_number_of_records_geq(uint64_t threshold) {
     if (threshold > get_histogram_max())
         return 0;
     size_t number_of_intervals = get_number_of_intervals();
     size_t i_index = get_interval_index(threshold);
     size_t sum = 0;
-    
+
     // parallelization: summation of counts
     #pragma omp parallel for reduction (+: sum)
     for (size_t i = i_index + 1; i < number_of_intervals; ++i) {
@@ -54,21 +54,21 @@ size_t Histogram::get_number_of_records_geq(unsigned int threshold) {
     return sum;
 }
 
-size_t Histogram::get_number_of_records_gt(unsigned int threshold) {
+size_t Histogram::get_number_of_records_gt(uint64_t threshold) {
     if (threshold == UINT_MAX)
         return 0;
     else
         return this->get_number_of_records_geq(threshold + 1);
 }
 
-size_t Histogram::get_number_of_records_leq(unsigned int threshold) {
+size_t Histogram::get_number_of_records_leq(uint64_t threshold) {
     if (threshold >= get_histogram_max()) {
         return get_total_number_of_records();
     }
     size_t number_of_intervals = get_number_of_intervals();
     size_t i_index = get_interval_index(threshold);
     size_t sum = 0;
-    
+
     // parallelization: summation of counts
     #pragma omp parallel for reduction (+: sum)
     for (size_t i = 0; i < i_index; ++i) {
@@ -83,15 +83,15 @@ size_t Histogram::get_number_of_records_leq(unsigned int threshold) {
     return sum;
 }
 
-size_t Histogram::get_number_of_records_lt(unsigned int threshold) {
+size_t Histogram::get_number_of_records_lt(uint64_t threshold) {
     if (threshold == 0)
         return 0;
     else
         return get_number_of_records_leq(threshold - 1);
 }
 
-size_t Histogram::get_number_of_records_geq_leq(unsigned int low, unsigned int high) {
-    
+size_t Histogram::get_number_of_records_geq_leq(uint64_t low, uint64_t high) {
+
     if (low > high)
         return 0;
     else if (low == 0)
@@ -103,7 +103,7 @@ size_t Histogram::get_number_of_records_geq_leq(unsigned int low, unsigned int h
     size_t left_i_index = get_interval_index(low);
     size_t right_i_index = get_interval_index(high);
     size_t sum = 0;
-    
+
     // parallelization: summation of counts
     #pragma omp parallel for reduction (+: sum)
     for (size_t i = left_i_index + 1; i < right_i_index; ++i) {
@@ -120,7 +120,7 @@ size_t Histogram::get_number_of_records_geq_leq(unsigned int low, unsigned int h
         size_t left_width = interval_width - (low % interval_width);
         size_t right_width = high % interval_width + 1;
         sum += static_cast<size_t>(static_cast<float>(interval_count[left_i_index])
-            * static_cast<float>(left_width) / float_width) + 
+            * static_cast<float>(left_width) / float_width) +
             static_cast<size_t>(static_cast<float>(interval_count[right_i_index])
             * static_cast<float>(right_width) / float_width);
     }
@@ -128,21 +128,21 @@ size_t Histogram::get_number_of_records_geq_leq(unsigned int low, unsigned int h
     return sum;
 }
 
-size_t Histogram::get_number_of_records_geq_lt(unsigned int low, unsigned int high) {
+size_t Histogram::get_number_of_records_geq_lt(uint64_t low, uint64_t high) {
     if (high == 0)
         return 0;
     else
         return get_number_of_records_geq_leq(low, high - 1);
 }
 
-size_t Histogram::get_number_of_records_gt_leq(unsigned int low, unsigned int high) {
+size_t Histogram::get_number_of_records_gt_leq(uint64_t low, uint64_t high) {
     if (low == UINT_MAX)
         return 0;
     else
         return get_number_of_records_geq_leq(low + 1, high);
 }
 
-size_t Histogram::get_number_of_records_gt_lt(unsigned int low, unsigned int high) {
+size_t Histogram::get_number_of_records_gt_lt(uint64_t low, uint64_t high) {
     if (low == UINT_MAX || high == 0)
         return 0;
     else
